Pawn Piece Transformation + Comments Structs.rs
I order to implement the feature that the player can choose a new piece when a pawn of his color reaches the opponents side of the board, I created a new window where the player clicks on the piece he wants. The type of the pawn will be changed to the selected one, the id will remain the same. Additionally, I commented out remaining methods in the structs.rs file.